fix: satisfy signed channel release gates
